Move store handling to `Pow.Plug.Base`
